requirements.txt
asgiref==3.8.1
azure-core==1.32.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.3
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.10
isodate==0.7.2
pillow==11.1.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.3.0
veryfi==4.0.1
pytz~=2024.2
babel~=2.16.0
